Introduce tab spaces replacing hard-coded spaces and quotation marks into the read mode output. Add a legend at the top of the output.
The rationale behind this change is to enable apetag output to be easily parsed by other tools. This is accomplished with tab-spaced output and with Bash and pipes, or with Windows and the command prompt. There are workarounds (CMD /F:OFF) for the Windows commandline to allow using a tab as a delimiter.